Structure tests by use case instead of one `happy_path` file
