+++ daemon-tests/src/lib.rs
@@ -1,10 +1,16 @@
+use crate::flow::is_next_offers_none;
+use crate::flow::next_maker_offers;
+use crate::flow::next_with;
+use crate::flow::one_cfd_with_state;
 use crate::mocks::monitor::MonitorActor;
 use crate::mocks::oracle::OracleActor;
 use crate::mocks::price_feed::PriceFeedActor;
 use crate::mocks::wallet::WalletActor;
+use anyhow::Context;
 use daemon::auto_rollover;
 use daemon::bdk::bitcoin::Amount;
 use daemon::bdk::bitcoin::Network;
+use daemon::bdk::bitcoin::SignedAmount;
 use daemon::bdk::bitcoin::Txid;
 use daemon::connection::connect;
 use daemon::connection::ConnectionStatus;
@@ -12,6 +18,7 @@ use daemon::libp2p_utils::create_connect_multiaddr;
 use daemon::maia_core::secp256k1_zkp::XOnlyPublicKey;
 use daemon::projection;
 use daemon::projection::Cfd;
+use daemon::projection::CfdState;
 use daemon::projection::Feeds;
 use daemon::projection::MakerOffers;
 use daemon::seed::RandomSeed;
@@ -19,6 +26,7 @@ use daemon::seed::Seed;
 use daemon::Environment;
 use daemon::HEARTBEAT_INTERVAL;
 use daemon::N_PAYOUTS;
+use maker::cfd::OfferParams;
 use model::libp2p::PeerId;
 use model::olivia::Announcement;
 use model::olivia::BitMexPriceEventId;
@@ -26,6 +34,7 @@ use model::CfdEvent;
 use model::CompleteFee;
 use model::Dlc;
 use model::EventKind;
+use model::FeeAccount;
 use model::FundingFee;
 use model::FundingRate;
 use model::Identity;
@@ -34,6 +43,7 @@ use model::OpeningFee;
 use model::OrderId;
 use model::Position;
 use model::Price;
+use model::Role;
 use model::TxFeeRate;
 use model::Usd;
 use model::SETTLEMENT_INTERVAL;
@@ -47,6 +57,7 @@ use std::time::Duration;
 use time::OffsetDateTime;
 use tokio::net::TcpListener;
 use tokio::sync::watch;
+use tokio_extras::time::sleep;
 use tokio_extras::Tasks;
 use tracing::instrument;
 use xtra::Actor;
@@ -57,6 +68,345 @@ use xtra_libp2p::multiaddress_ext::MultiaddrExt;
 pub mod flow;
 pub mod maia;
 pub mod mocks;
+
+#[macro_export]
+macro_rules! confirm {
+    (lock transaction, $id:expr, $maker:expr, $taker:expr) => {
+        $maker
+            .mocks
+            .monitor()
+            .await
+            .confirm_lock_transaction($id)
+            .await;
+        $taker
+            .mocks
+            .monitor()
+            .await
+            .confirm_lock_transaction($id)
+            .await;
+    };
+    (commit transaction, $id:expr, $maker:expr, $taker:expr) => {
+        $maker
+            .mocks
+            .monitor()
+            .await
+            .confirm_commit_transaction($id)
+            .await;
+        $taker
+            .mocks
+            .monitor()
+            .await
+            .confirm_commit_transaction($id)
+            .await;
+    };
+    (refund transaction, $id:expr, $maker:expr, $taker:expr) => {
+        $maker
+            .mocks
+            .monitor()
+            .await
+            .confirm_refund_transaction($id)
+            .await;
+        $taker
+            .mocks
+            .monitor()
+            .await
+            .confirm_refund_transaction($id)
+            .await;
+    };
+    (close transaction, $id:expr, $maker:expr, $taker:expr) => {
+        $maker
+            .mocks
+            .monitor()
+            .await
+            .confirm_close_transaction($id)
+            .await;
+        $taker
+            .mocks
+            .monitor()
+            .await
+            .confirm_close_transaction($id)
+            .await;
+    };
+    (cet, $id:expr, $maker:expr, $taker:expr) => {
+        $maker.mocks.monitor().await.confirm_cet($id).await;
+        $taker.mocks.monitor().await.confirm_cet($id).await;
+    };
+}
+
+#[macro_export]
+macro_rules! expire {
+    (cet timelock, $id:expr, $maker:expr, $taker:expr) => {
+        $maker.mocks.monitor().await.expire_cet_timelock($id).await;
+        $taker.mocks.monitor().await.expire_cet_timelock($id).await;
+    };
+    (refund timelock, $id:expr, $maker:expr, $taker:expr) => {
+        $maker
+            .mocks
+            .monitor()
+            .await
+            .expire_refund_timelock($id)
+            .await;
+        $taker
+            .mocks
+            .monitor()
+            .await
+            .expire_refund_timelock($id)
+            .await;
+    };
+}
+
+/// Simulate oracle attestation for both actor systems
+#[macro_export]
+macro_rules! simulate_attestation {
+    ($maker:expr, $taker:expr, $order_id:expr, $attestation:expr) => {{
+        tracing::debug!("Simulating attestation: {:?}", $attestation);
+
+        $maker
+            .mocks
+            .oracle()
+            .await
+            .simulate_attestation($order_id, $attestation)
+            .await;
+
+        $taker
+            .mocks
+            .oracle()
+            .await
+            .simulate_attestation($order_id, $attestation)
+            .await;
+    }};
+}
+
+/// Waits until the CFDs for both maker and taker are in the given state.
+#[macro_export]
+macro_rules! wait_next_state {
+    ($id:expr, $maker:expr, $taker:expr, $maker_state:expr, $taker_state:expr) => {
+        let wait_until_taker = next_with($taker.cfd_feed(), |maybe_cfds| {
+            maybe_cfds.and_then(one_cfd_with_state($taker_state))
+        });
+        let wait_until_maker = next_with($maker.cfd_feed(), |maybe_cfds| {
+            maybe_cfds.and_then(one_cfd_with_state($maker_state))
+        });
+
+        let (taker_cfd, maker_cfd) = tokio::join!(wait_until_taker, wait_until_maker);
+        let taker_cfd = taker_cfd.unwrap();
+        let maker_cfd = maker_cfd.unwrap();
+
+        assert_eq!(
+            taker_cfd.order_id, maker_cfd.order_id,
+            "order id mismatch between maker and taker"
+        );
+        assert_eq!(taker_cfd.order_id, $id, "unexpected order id in the taker");
+        assert_eq!(maker_cfd.order_id, $id, "unexpected order id in the maker");
+    };
+    ($id:expr, $maker:expr, $taker:expr, $state:expr) => {
+        wait_next_state!($id, $maker, $taker, $state, $state)
+    };
+}
+
+/// Hide the implementation detail of arriving at the Cfd open state.
+/// Useful when reading tests that should start at this point.
+/// For convenience, returns also OrderId of the opened Cfd.
+/// `announcement` is used during Cfd's creation.
+pub async fn start_from_open_cfd_state(
+    announcement: Announcement,
+    position_maker: Position,
+) -> (Maker, Taker, OrderId, FeeStructure) {
+    let mut maker = Maker::start(&MakerConfig::default()).await;
+    let mut taker = Taker::start(
+        &TakerConfig::default(),
+        maker.listen_addr,
+        maker.identity,
+        maker.connect_addr.clone(),
+    )
+    .await;
+
+    is_next_offers_none(taker.offers_feed()).await.unwrap();
+
+    let offer_params = dummy_offer_params(position_maker);
+
+    let quantity = Usd::new(dec!(100));
+    let taker_leverage = Leverage::TWO;
+
+    let fee_structure = FeeStructure::new(
+        offer_params.clone(),
+        quantity,
+        taker_leverage,
+        position_maker,
+    );
+
+    maker.set_offer_params(offer_params).await;
+
+    let (_, received) = next_maker_offers(maker.offers_feed(), taker.offers_feed())
+        .await
+        .unwrap();
+
+    taker
+        .mocks
+        .mock_oracle_announcement_with(announcement.clone())
+        .await;
+    maker
+        .mocks
+        .mock_oracle_announcement_with(announcement)
+        .await;
+
+    let order_to_take = match position_maker {
+        Position::Short => received.short,
+        Position::Long => received.long,
+    }
+    .context("Order for expected position not set")
+    .unwrap();
+
+    taker
+        .system
+        .take_offer(order_to_take.id, quantity, taker_leverage)
+        .await
+        .unwrap();
+
+    wait_next_state!(order_to_take.id, maker, taker, CfdState::PendingSetup);
+
+    maker.mocks.mock_party_params().await;
+    taker.mocks.mock_party_params().await;
+
+    maker.mocks.mock_wallet_sign_and_broadcast().await;
+    taker.mocks.mock_wallet_sign_and_broadcast().await;
+
+    maker.system.accept_order(order_to_take.id).await.unwrap();
+    wait_next_state!(order_to_take.id, maker, taker, CfdState::ContractSetup);
+
+    sleep(Duration::from_secs(5)).await; // need to wait a bit until both transition
+    wait_next_state!(order_to_take.id, maker, taker, CfdState::PendingOpen);
+
+    confirm!(lock transaction, order_to_take.id, maker, taker);
+    wait_next_state!(order_to_take.id, maker, taker, CfdState::Open);
+
+    (maker, taker, order_to_take.id, fee_structure)
+}
+
+pub struct FeeStructure {
+    /// Opening fee charged by the maker
+    opening_fee: OpeningFee,
+
+    /// Funding fee for the first 24h calculated when opening a Cfd
+    initial_funding_fee: FundingFee,
+
+    /// The maker's position for the Cfd
+    maker_position: Position,
+
+    offer_params: OfferParams,
+    quantity: Usd,
+    taker_leverage: Leverage,
+}
+
+impl FeeStructure {
+    pub fn new(
+        offer_params: OfferParams,
+        quantity: Usd,
+        taker_leverage: Leverage,
+        maker_position: Position,
+    ) -> Self {
+        let initial_funding_fee = match maker_position {
+            Position::Long => FundingFee::calculate(
+                offer_params.price_long.unwrap(),
+                quantity,
+                Leverage::ONE,
+                taker_leverage,
+                offer_params.funding_rate_long,
+                SETTLEMENT_INTERVAL.whole_hours(),
+            )
+            .unwrap(),
+            Position::Short => FundingFee::calculate(
+                offer_params.price_short.unwrap(),
+                quantity,
+                taker_leverage,
+                Leverage::ONE,
+                offer_params.funding_rate_short,
+                SETTLEMENT_INTERVAL.whole_hours(),
+            )
+            .unwrap(),
+        };
+
+        Self {
+            opening_fee: offer_params.opening_fee,
+            initial_funding_fee,
+            maker_position,
+            offer_params,
+            quantity,
+            taker_leverage,
+        }
+    }
+
+    pub fn predict_fees(
+        &self,
+        accumulated_rollover_hours_to_charge: i64,
+    ) -> (SignedAmount, SignedAmount) {
+        if accumulated_rollover_hours_to_charge == 0 {
+            tracing::info!("Predicting fees before first rollover")
+        } else {
+            tracing::info!(
+                "Predicting fee for {} hours",
+                accumulated_rollover_hours_to_charge
+            );
+        }
+
+        tracing::debug!("Opening fee: {}", self.opening_fee);
+
+        let mut maker_fee_account = FeeAccount::new(self.maker_position, Role::Maker)
+            .add_opening_fee(self.opening_fee)
+            .add_funding_fee(self.initial_funding_fee);
+
+        let taker_position = self.maker_position.counter_position();
+        let mut taker_fee_account = FeeAccount::new(taker_position, Role::Taker)
+            .add_opening_fee(self.opening_fee)
+            .add_funding_fee(self.initial_funding_fee);
+
+        tracing::debug!(
+            "Maker fees including opening and initial funding fee: {}",
+            maker_fee_account.balance()
+        );
+
+        tracing::debug!(
+            "Taker fees including opening and initial funding fee: {}",
+            taker_fee_account.balance()
+        );
+
+        let accumulated_hours_to_charge = match self.maker_position {
+            Position::Long => FundingFee::calculate(
+                self.offer_params.price_long.unwrap(),
+                self.quantity,
+                Leverage::ONE,
+                self.taker_leverage,
+                self.offer_params.funding_rate_long,
+                accumulated_rollover_hours_to_charge,
+            )
+            .unwrap(),
+            Position::Short => FundingFee::calculate(
+                self.offer_params.price_short.unwrap(),
+                self.quantity,
+                self.taker_leverage,
+                Leverage::ONE,
+                self.offer_params.funding_rate_short,
+                accumulated_rollover_hours_to_charge,
+            )
+            .unwrap(),
+        };
+
+        maker_fee_account = maker_fee_account.add_funding_fee(accumulated_hours_to_charge);
+        taker_fee_account = taker_fee_account.add_funding_fee(accumulated_hours_to_charge);
+
+        tracing::debug!(
+            "Maker fees including all fees: {}",
+            maker_fee_account.balance()
+        );
+
+        tracing::debug!(
+            "Taker fees including all fees: {}",
+            taker_fee_account.balance()
+        );
+
+        (maker_fee_account.balance(), taker_fee_account.balance())
+    }
+}
 
 fn oracle_pk() -> XOnlyPublicKey {
     XOnlyPublicKey::from_str("ddd4636845a90185991826be5a494cde9f4a6947b1727217afedc6292fa4caf7")
@@ -268,8 +618,8 @@ impl Maker {
         }
     }
 
-    pub async fn set_offer_params(&mut self, offer_params: maker::cfd::OfferParams) {
-        let maker::cfd::OfferParams {
+    pub async fn set_offer_params(&mut self, offer_params: OfferParams) {
+        let OfferParams {
             price_long,
             price_short,
             min_quantity,
@@ -513,55 +863,6 @@ impl Taker {
     }
 }
 
-/// Simulate oracle attestation for both actor systems
-#[macro_export]
-macro_rules! simulate_attestation {
-    ($maker:expr, $taker:expr, $order_id:expr, $attestation:expr) => {{
-        tracing::debug!("Simulating attestation: {:?}", $attestation);
-
-        $maker
-            .mocks
-            .oracle()
-            .await
-            .simulate_attestation($order_id, $attestation)
-            .await;
-
-        $taker
-            .mocks
-            .oracle()
-            .await
-            .simulate_attestation($order_id, $attestation)
-            .await;
-    }};
-}
-
-/// Waits until the CFDs for both maker and taker are in the given state.
-#[macro_export]
-macro_rules! wait_next_state {
-    ($id:expr, $maker:expr, $taker:expr, $maker_state:expr, $taker_state:expr) => {
-        let wait_until_taker = next_with($taker.cfd_feed(), |maybe_cfds| {
-            maybe_cfds.and_then(one_cfd_with_state($taker_state))
-        });
-        let wait_until_maker = next_with($maker.cfd_feed(), |maybe_cfds| {
-            maybe_cfds.and_then(one_cfd_with_state($maker_state))
-        });
-
-        let (taker_cfd, maker_cfd) = tokio::join!(wait_until_taker, wait_until_maker);
-        let taker_cfd = taker_cfd.unwrap();
-        let maker_cfd = maker_cfd.unwrap();
-
-        assert_eq!(
-            taker_cfd.order_id, maker_cfd.order_id,
-            "order id mismatch between maker and taker"
-        );
-        assert_eq!(taker_cfd.order_id, $id, "unexpected order id in the taker");
-        assert_eq!(maker_cfd.order_id, $id, "unexpected order id in the maker");
-    };
-    ($id:expr, $maker:expr, $taker:expr, $state:expr) => {
-        wait_next_state!($id, $maker, $taker, $state, $state)
-    };
-}
-
 pub fn dummy_quote() -> Quote {
     Quote {
         timestamp: OffsetDateTime::now_utc(),
@@ -571,13 +872,13 @@ pub fn dummy_quote() -> Quote {
 }
 
 // Offer params allowing a single position, either short or long
-pub fn dummy_offer_params(position_maker: Position) -> maker::cfd::OfferParams {
+pub fn dummy_offer_params(position_maker: Position) -> OfferParams {
     let (price_long, price_short) = match position_maker {
         Position::Long => (Some(Price::new(dummy_price()).unwrap()), None),
         Position::Short => (None, Some(Price::new(dummy_price()).unwrap())),
     };
 
-    maker::cfd::OfferParams {
+    OfferParams {
         price_long,
         price_short,
         min_quantity: Usd::new(dec!(100)),
